Changed structure
Plot-functions separated from each other in individual files, easier.
plot_bar to plot a bar-chart
plot_line to plot a line-graph
plot_pie for a pie-chart
plot_world for the world map
sentiment_plot is an example how to use it
